sửa chức năng xoá

+++ NguyenHuynhTuong65134116_fitCMS/src/main/java/thiGK/ntu65134116/NguyenHuynhTuong65134116_fitCMS/services/PageService.java
@@ -26,6 +26,9 @@ public class PageService {
 	}
 	
 	public void deletePage(int id) {
+	    if (!pRepos.existsById(id)) {
+	        throw new RuntimeException("ID không tồn tại");
+	    }
 	    pRepos.deleteById(id);
 	}
 }
